promote import TimeWindowSelectionContext
