+++ src/components/TimeWindowSelectionContext.tsx
@@ -1,0 +1,17 @@
+/**
+ *  TimeWindowSelectionContext.ts
+ *
+ *  @copyright 2025 Digital Aid Seattle
+ *
+ */
+import { createContext } from "react";
+
+interface TimeWindowContextType {
+    selection: string[],
+    setSelection: React.Dispatch<React.SetStateAction<string[]>>
+}
+
+export const TimeWindowSelectionContext = createContext<TimeWindowContextType>({
+    selection: [] as string[],
+    setSelection: () => []
+})
